IPTV/HLS/M3U integer overflow
Fix integer overflow in reading IPTV/HLS segments by computing the
segment size times the number of segments now in a 64 bit integer.
This fixes playback of the Spanish channel RTVE.es
on https://rtvelivestream.akamaized.net/rtvesec/24h/24h_main_dvr_720.m3u8

Refs #936

+++ mythtv/libs/libmythtv/recorders/HLS/HLSReader.cpp
@@ -988,7 +988,7 @@ int HLSReader::DownloadSegmentData(MythSingleDownload& downloader,
     }
 #endif
 
-    int segment_len = buffer.size();
+    int64_t segment_len = buffer.size();
 
     m_bufLock.lock();
     if (m_buffer.size() > segment_len * playlist_size)
